+++ out/provider/vision/index.js
@@ -1,9 +1,10 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.resolveImageMessages = exports.setVisionProxyModel = exports.createVisionModelGetter = void 0;
+exports.resolveImageMessages = exports.setVisionProxyModel = exports.createVisionModelGetter = exports.describeViaEndpoint = void 0;
 var model_1 = require("./model");
 Object.defineProperty(exports, "createVisionModelGetter", { enumerable: true, get: function () { return model_1.createVisionModelGetter; } });
 Object.defineProperty(exports, "setVisionProxyModel", { enumerable: true, get: function () { return model_1.setVisionProxyModel; } });
+Object.defineProperty(exports, "describeViaEndpoint", { enumerable: true, get: function () { return model_1.describeViaEndpoint; } });
 var resolve_1 = require("./resolve");
 Object.defineProperty(exports, "resolveImageMessages", { enumerable: true, get: function () { return resolve_1.resolveImageMessages; } });
 //# sourceMappingURL=index.js.map